Add connection check and fix eeprom write issue

+++ Core/Inc/EEPROM.h
@@ -32,8 +32,8 @@ typedef struct
 void EEPROM_Init(I2C_HandleTypeDef *hi2c);
 void EEPROM_Set_Parameters(uint8_t mem_address, uint16_t mem_size, uint16_t page_size);
 
-STD_Error EEPROM_Write(uint8_t address, const uint8_t * data, size_t size);
-STD_Error EEPROM_Read(uint8_t address, uint8_t *data, size_t size);
+STD_Error EEPROM_Write(uint8_t address, const uint8_t * data, int size);
+STD_Error EEPROM_Read(uint8_t address, uint8_t *data, int size);
 
 EEPROM_Config_T *EEPROM_Get_ConfigData(void);
